Fix XReqLib elaboration problems for good !!!
What was the problem:

XReqLib was compiled two times, once for Ada clients and another time for
other languages (that features auto-elaboration). The problem is that ALI
files were installed at the same place. And the Ada programs got ALI files
from the auto-elaborated library. But their binary was not auto
elaborated. So, the binder did not generate elaboration for XReqLib.

Solution: diossociate all directories

Useful gprbuild flags:
-v:      show complete commands executed
-dn:     don't delete GNAT-TEMP files
-b:      bind only
-bargs:  arguments to the binder
...